pyproject.toml
[tool.poetry]
name = "beam-split-paragraph"
version = "0.1.0"
description = ""
authors = ["kevin-weitgenant <kevin.weitgenant@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

beam-client = "^0.2.112"
wtpsplit = "^2.1.1"
torch = "^2.5.1"
onnxruntime-gpu = "1.19.2"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
